extract accommodation
